* main: (652 commits)
  prune dead files
  punctuation in clinic name causes consultation request to crash
  no arigato mr roboto
  update user rights in fax cover page to just the fax object
  gitignore
  reroute fax account selection from consultation request
  use bean copy instead of null value to sanitize sensitive fax passwords
  add empty and null variable checks to new fax password encryption process
  set default letterhead and referring providers on load consult request
  Fix NullPointerException in BillingReprocessBillAction(s)
  add missing method signature to PropertyDao that caused BC Billing to fail loading.
  update look and feel in patient independent eform list
  CreateLab.jsp bootstrap html5 jquery 3.6 autocomplete and minimum validations required/possible to create a lab that displays
  Fix demographic search to include inactive patients for document linking in Inbox
  MSP eligibility check: eliminate prototype, fix for mobile devices, fix race condition
  Fix: Correct parameter order in sanitizeElementByPattern function calling
  fix custom fax number override default fax number in consult request
  Enhance encryption/decryption functionality with AES/GCM for secure storage
  correct data constraints in the relationship model
  Update ExcellerisON HL7 lab presentation and PDF report per LifeLabs ON guidelines:
  ...

# Conflicts:
#	src/main/resources/clientKeystore.properties

+++ src/main/java/org/oscarehr/provider/web/ProviderPropertyAction.java
@@ -24,18 +24,8 @@
 
 package org.oscarehr.provider.web;
 
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Hashtable;
-import java.util.List;
-import java.util.Vector;
-
-import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletResponse;
-
 import org.apache.commons.beanutils.BeanComparator;
+import org.apache.commons.lang3.StringUtils;
 import org.apache.struts.action.ActionForm;
 import org.apache.struts.action.ActionForward;
 import org.apache.struts.action.ActionMapping;
@@ -52,10 +42,13 @@ import org.oscarehr.common.model.UserProperty;
 import org.oscarehr.util.LoggedInInfo;
 import org.oscarehr.util.MiscUtils;
 import org.oscarehr.util.SpringUtils;
-
 import oscar.eform.EFormUtil;
 import oscar.log.LogAction;
 import oscar.oscarEncounter.oscarConsultationRequest.pageUtil.EctConsultationFormRequestUtil;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+import java.util.*;
 
 /**
  *
@@ -153,6 +146,31 @@ public class ProviderPropertyAction extends DispatchAction {
 
          request.setAttribute("status", "success");
          return actionmapping.findForward("success");
+    }
+
+    /**
+     * typically set from inside the JSP class providerupdatepreference.jsp
+     * @param request
+     */
+    public static void updateOrCreateProviderProperties(HttpServletRequest request) {
+        LoggedInInfo loggedInInfo = LoggedInInfo.getLoggedInInfoFromSession(request);
+        UserPropertyDAO propertyDAO = SpringUtils.getBean(UserPropertyDAO.class);
+        String providerNo = loggedInInfo.getLoggedInProviderNo();
+
+        List<UserProperty> userProperties = new ArrayList<>();
+        String propertyValue;
+        UserProperty property;
+
+        propertyValue = StringUtils.trimToNull(request.getParameter(UserProperty.SCHEDULE_WEEK_VIEW_WEEKENDS));
+        property = propertyDAO.getProp(providerNo, UserProperty.SCHEDULE_WEEK_VIEW_WEEKENDS);
+        if (property == null) {
+            property = new UserProperty();
+            property.setProviderNo(providerNo);
+            property.setName(UserProperty.SCHEDULE_WEEK_VIEW_WEEKENDS);
+        }
+        property.setValue(String.valueOf(Boolean.parseBoolean(propertyValue)));
+        propertyDAO.saveProp(property);
+
     }
     public ActionForward viewDefaultSex(ActionMapping actionmapping,
                                ActionForm actionform,
@@ -415,7 +433,7 @@ public class ProviderPropertyAction extends DispatchAction {
         }
         if(mode.equals("new")){
             //save and get most recent id
-            QueueDao queueDao = (QueueDao) SpringUtils.getBean("queueDao");
+            QueueDao queueDao = (QueueDao) SpringUtils.getBean(QueueDao.class);
             queueDao.addNewQueue(defaultQ);
             String lastId=queueDao.getLastId();
             prop.setValue(lastId);
@@ -447,7 +465,7 @@ public class ProviderPropertyAction extends DispatchAction {
         if(prop==null){
             prop=new UserProperty();
         }
-        QueueDao queueDao = (QueueDao) SpringUtils.getBean("queueDao");
+        QueueDao queueDao = (QueueDao) SpringUtils.getBean(QueueDao.class);
         List<Hashtable> queues= queueDao.getQueues();
         Collection<LabelValueBean> viewChoices=new ArrayList<LabelValueBean>();
         viewChoices.add(new LabelValueBean("None","-1"));
@@ -2048,12 +2066,22 @@ public ActionForward viewEDocBrowserInDocumentReport(ActionMapping actionmapping
 		property.setName(UserProperty.TICKLER_TASK_ASSIGNEE);
 	}
 
-	if(delete){
-	 userPropertyDAO.delete(property);
-	}else{
-	 property.setValue(tickerTaskAssignee);
-	 userPropertyDAO.saveProp(property);
-	}
+	try {
+        if (delete) {
+            if (property.getId() != null) {
+                userPropertyDAO.delete(property);
+            }
+        } else {
+            property.setValue(tickerTaskAssignee);
+            userPropertyDAO.saveProp(property);
+        }
+    } catch (Exception e) {
+        // Return to the success page even though the pereference is not changed from default
+        // Avoid the error displays
+        request.setAttribute("status", "success");
+        return actionmapping.findForward("complete");
+    }
+    
 
 	    request.setAttribute("status", "success");
